+++ .gitignore
@@ -1,2 +1,3 @@
 /chroma_db_dir
 /venv
+.devcontainer
